Task 2: Modular compilation
